Scale training to full dual-H200 capacity
Four changes to push GPU utilization from 24% to near-100%:

1. Shield LSTM to float32 (model.py, actor_critic.py)
   Cast input to float32 before the LSTM cell so BF16 autocast is safe
   for the linear head layers without causing NaN in hidden states.

2. Enable BF16 AMP + GradScaler (trainer.py, ppo_trainer.py)
   H200 BF16 tensor cores deliver ~2x throughput vs FP32.  GradScaler
   is a transparent no-op on non-CUDA devices.

3. Fix DataParallel bypass bug (ppo_trainer.py)
   _ppo_update was calling model.module.forward_full() directly, routing
   all gradient computation through GPU 0 only.  ActorCriticLSTM.forward
   now returns all three heads so self.model(x) works through DataParallel.

4. Vectorize PPO rollout collection (ppo_trainer.py)
   Previous: rollout_days × bars-per-day individual batch-1 forward passes.
   Now: all active episodes stepped in lock-step, observations stacked into
   a single (n_active, seq_len, features) tensor per bar — one GPU call
   feeds both H200s instead of thousands of tiny dispatches.

Also: auto-scale batch_size and minibatch_size by GPU count, auto-enable
torch.compile on H200, scale num_workers to CPU count, add prefetch_factor=2.

+++ backtesting/ml/model.py
@@ -58,7 +58,10 @@ class LSTMModel(nn.Module):
         x : (batch, seq_len, n_features)
         returns logits : (batch, n_classes)
         """
-        out, _ = self.lstm(x)        # (batch, seq_len, hidden)
+        # LSTM must run in float32 — BF16 sequential accumulation loses
+        # mantissa precision and produces NaN in hidden states.
+        # The outer autocast context covers the linear head layers.
+        out, _ = self.lstm(x.float())   # (batch, seq_len, hidden)
         h = out[:, -1, :]            # last timestep
         h = self.norm(h)
         return self.head(h)          # (batch, n_classes)
